// backend/src/resources/users/user/enums/user.enum.ts
export enum USER_STATUS {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
  EMAIL_VERIFICATION_PENDING = 'email verification pending',
}

export enum CREATED_BY {
  ADMIN = 'admin',
  USER = 'signup',
}

export enum USER_NAME_OPTIONS {
  NAME = 'name',
  PRICE = 'contactNumber',
  CITY = 'city',
  STATE = 'state',
  STATUS = 'status',
}